Refactor e2e tests to split per language
