+++ e2e-tests/testdata/echo_client/golang/client/goclient.go
