+++ e2e-tests/testdata/echo_client/js/client.js
